🔧 Fix Dropdown errors
